+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - set the correct name of the `RegisterFailureReport` service method used in distributed mode (#485)
+- Return the correct models in `GetComputeTaskInputModels` for composite tasks (#499)
 
 ## [0.5.0] - 2022-01-16